+++ core-java-modules/core-java-numbers-conversions-2/README.md
@@ -1,3 +1,9 @@
 ### Relevant Articles:
-- [Convert From int to short in Java](https://www.baeldung.com/java-int-short-conversion)
 - [Convert int to Unsigned byte in Java](https://www.baeldung.com/java-convert-int-unsigned-byte)
+- [Convert a Number to a Letter in Java](https://www.baeldung.com/java-convert-number-to-letter)
+- [Convert Long to BigDecimal in Java](https://www.baeldung.com/java-convert-long-bigdecimal)
+- [Converting from float to BigDecimal in Java](https://www.baeldung.com/java-convert-float-bigdecimal)
+- [Rounding Up a Number to Nearest Multiple of 5 in Java](https://www.baeldung.com/java-round-nearest-multiple-five)
+- [Convert byte to int Type in Java](https://www.baeldung.com/java-byte-integer-conversion)
+- [Converting BigDecimal to Integer in Java](https://www.baeldung.com/java-integer-bigdecimal-conversion)
+- [Converting Integer to BigDecimal in Java](https://www.baeldung.com/java-from-integer-to-bigdecimal)
