+++ core-java-modules/core-java-numbers-conversions-2/src/test/java/com/baeldung/bigdecimaltointeger/SmallBigDecimalConversionArgumentsProvider.java
